+++ src/components/Box/index.ts
@@ -1,0 +1,5 @@
+import RvBox from "./RvBox";
+import RvFlex from "./RvFlex";
+
+export { RvBox, RvFlex };
+export * from "./style";