+++ src/components/Div/Flex.tsx
@@ -1,9 +1,0 @@
-import { FC } from "react";
-import type { FlexFCType } from "../../types/components/Div";
-import { StyledFlex } from "./style";
-
-const Flex: FC<FlexFCType> = ({ children, ...props }) => {
-	return <StyledFlex {...props}>{children}</StyledFlex>;
-};
-
-export default Flex;