+++ src/components/Div/index.ts
@@ -1,5 +1,0 @@
-import Box from "./Box";
-import Flex from "./Flex";
-
-export { Box, Flex };
-export * from "./style";