+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export * from "./Div";
+export * from "./Box";
 export * from "./Button";
 export * from "./config";